doorf4_SeqFn: drop outer (u8) on GameBit_Get^val xor drops redundant clrlwi (98.07->98.60)
Both xor operands are already (u8)-masked, so the result fits in u8 and the
outer (u8) cast only produced an extra clrlwi r4,r4,24 after each xor.
GameBit_Set takes an int value param, so dropping the outer cast is faithful.
Matches target's mask-each-operand-then-xor shape (#37/#74).

(cherry picked from commit 0af8c42b5e3d50d15831f55b75c050627c5f63ca)

+++ src/main/dll/dll_00F4_doorf4.c
@@ -454,13 +454,13 @@ int doorf4_SeqFn(int* obj, int unused, ObjAnimUpdateState* animUpdate)
                     if (*(s16*)(def + 0x20) != -1)
                     {
                         GameBit_Set(*(s16*)(def + 0x20),
-                                    (u8)((u8)GameBit_Get(*(s16*)(def + 0x20)) ^ (u8) * (s16*)(def + 0x1c)));
+                                    (u8)GameBit_Get(*(s16*)(def + 0x20)) ^ (u8) * (s16*)(def + 0x1c));
                     }
                 }
                 else if (*(s16*)(def + 0x1a) != -1)
                 {
                     GameBit_Set(*(s16*)(def + 0x1a),
-                                (u8)((u8)GameBit_Get(*(s16*)(def + 0x1a)) ^ (u8)(*(s16*)(def + 0x1c) >> 8)));
+                                (u8)GameBit_Get(*(s16*)(def + 0x1a)) ^ (u8)(*(s16*)(def + 0x1c) >> 8));
                 }
                 if (sd <= lbl_803E3648)
                 {
@@ -518,13 +518,13 @@ int doorf4_SeqFn(int* obj, int unused, ObjAnimUpdateState* animUpdate)
                     if (*(s16*)(def + 0x20) != -1)
                     {
                         GameBit_Set(*(s16*)(def + 0x20),
-                                    (u8)((u8)GameBit_Get(*(s16*)(def + 0x20)) ^ (u8) * (s16*)(def + 0x1c)));
+                                    (u8)GameBit_Get(*(s16*)(def + 0x20)) ^ (u8) * (s16*)(def + 0x1c));
                     }
                 }
                 else if (*(s16*)(def + 0x1a) != -1)
                 {
                     GameBit_Set(*(s16*)(def + 0x1a),
-                                (u8)((u8)GameBit_Get(*(s16*)(def + 0x1a)) ^ (u8)(*(s16*)(def + 0x1c) >> 8)));
+                                (u8)GameBit_Get(*(s16*)(def + 0x1a)) ^ (u8)(*(s16*)(def + 0x1c) >> 8));
                 }
                 switch (((GameObject*)obj)->anim.seqId)
                 {
